Move KProve into its own submodule
